requirements.txt
# case insensitive
django>=1.8,<2.0
psycopg2
djangorestframework
django-rest-swagger
